[Android/Feat] 워치 api 연동 끝 S12P31D103-206

+++ Android/domain/src/main/java/com/hogumiwarts/domain/repository/LightRepository.kt
@@ -9,9 +9,9 @@ import com.hogumiwarts.domain.model.light.LightTemperatureResult
 
 interface LightRepository {
 
-    suspend fun getLightStatus(deviceId: Int): GetLightStatusResult
+    suspend fun getLightStatus(deviceId: Long): GetLightStatusResult
 
-    suspend fun patchLightPower(deviceId: Int, activated: Boolean): PatchSwitchPowerResult
+    suspend fun patchLightPower(deviceId: Long, activated: Boolean): PatchSwitchPowerResult
 
     suspend fun patchLightBright(deviceId: Long, brightness: Int): LightBrightResult
