+++ Android/wearable/src/main/java/com/hogumiwarts/lumos/presentation/ui/screens/control/light/LightIntent.kt
@@ -9,6 +9,6 @@ sealed class LightIntent {
 
     data class ChangeLightBright(val deviceId: Long, val brightness: Int): LightIntent()
 
-    data class ChangeLightColor(val deviceId: Long, val color: Int): LightIntent()
+    data class ChangeLightColor(val deviceId: Long, val color: Float): LightIntent()
 
 }
